Добавлена реализация класса Environment. Тип Int16 заменён на Int32 во всей системе.

+++ Project/Rybocompleks.GUI/Rybocompleks.Data/Classes/LightMeasurment.cs
@@ -19,7 +19,7 @@ namespace Rybocompleks.Data.Classes
             return TurnedOn;
         }
 
-        public Int16 GetPropertyID()
+        public Int32 GetPropertyID()
         {
             throw new NotImplementedException();
         }
